removed gift, will have to use variation

+++ order/models.py
@@ -42,7 +42,7 @@ class Order(models.Model):
     order_note = models.CharField(max_length=100, blank=True)
     order_total = models.FloatField()
     tax = models.FloatField()
-    delivery_charge = models.FloatField(null=True)
+    delivery_charge = models.SmallIntegerField(null=True)
     status = models.CharField(max_length=10, choices=STATUS, default='New')
     ip = models.CharField(blank=True, max_length=20)
     is_ordered = models.BooleanField(default=False)
